requirements.txt
Django==4.2.18
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
drf-yasg==1.21.8